Extracted magic energy
