+++ planet/Objects.ocd/Libraries.ocd/MagicEnergy.ocd/DefCore.txt
@@ -1,0 +1,5 @@
+[DefCore]
+id=Library_MagicEnergy
+Version=8,0
+Category=C4D_StaticBack
+HideInCreator=true
